+++ packages/snaps-rpc-methods/jest.config.js
@@ -10,10 +10,10 @@ module.exports = deepmerge(baseConfig, {
   ],
   coverageThreshold: {
     global: {
-      branches: 97.38,
+      branches: 97.4,
       functions: 98.92,
       lines: 99.22,
-      statements: 98.95,
+      statements: 98.96,
     },
   },
 });